backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
psycopg2-binary==2.9.9
python-multipart==0.0.12
requests==2.32.3
google-genai>=0.5.0
pillow==10.4.0
python-dotenv==1.0.1
